divided dtos and business objects
