+++ src/typings/dto.d.ts
@@ -1,12 +1,6 @@
 declare module 'app/dto' {
-  interface IBusinessObject {
-    id: number
-  }
 
-  type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>
-  type WithoutId<T extends IBusinessObject> = Omit<T, keyof IBusinessObject>
-
-  interface ITimestampDto extends IBusinessObject {
+  interface ITimestampDto {
     id: number
     taskId: number
     datetimeStart: string
@@ -14,24 +8,14 @@ declare module 'app/dto' {
     comment: string
   }
 
-  interface ITimestamp extends IBusinessObject {
-    id: number
-    taskId: number
-    datetimeStart: Date
-    datetimeEnd?: Date
-    comment: string
-  }
-
-  interface ITask extends IBusinessObject {
+  interface ITaskDto {
     id: number
     projectId: number
     title: string
   }
 
-  interface IProject extends IBusinessObject {
+  interface IProjectDto {
     id: number
     title: string
   }
-
-  type Dictionary<T> = Map<number, T>
 }
